requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via starlette
attrs==23.1.0
    # via
    #   outcome
    #   trio
certifi==2023.7.22
    # via selenium
click==8.1.6
    # via uvicorn
exceptiongroup==1.1.2
    # via trio-websocket
fastapi==0.100.1
    # via -r requirements.in
h11==0.14.0
    # via
    #   uvicorn
    #   wsproto
idna==3.4
    # via
    #   anyio
    #   trio
outcome==1.2.0
    # via trio
pydantic==2.1.1
    # via fastapi
pydantic-core==2.4.0
    # via pydantic
pysocks==1.7.1
    # via urllib3
python-decouple==3.8
    # via -r requirements.in
selenium==4.10.0
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   trio
sortedcontainers==2.4.0
    # via trio
starlette==0.27.0
    # via fastapi
tenacity==8.2.2
    # via -r requirements.in
trio==0.22.2
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.10.3
    # via selenium
typing-extensions==4.7.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3[socks]==2.0.4
    # via selenium
uvicorn==0.23.1
    # via -r requirements.in
wsproto==1.2.0
    # via trio-websocket
